splitted canvas utils in two files, one generic and one for shape
